chore: Rename lib/file.js to lib/file-coverage.js
