+++ packages/istanbul-lib-coverage/lib/file-coverage.js
